+++ objc2/examples/nspasteboard.rs
@@ -9,7 +9,7 @@ use std::mem::ManuallyDrop;
 use objc2::foundation::{NSArray, NSDictionary, NSInteger, NSObject, NSString};
 use objc2::rc::{Id, Shared};
 use objc2::runtime::{Class, Object};
-use objc2::{extern_class, msg_send, msg_send_bool, msg_send_id, ClassType};
+use objc2::{extern_class, msg_send, msg_send_id, ClassType};
 
 type NSPasteboardType = NSString;
 type NSPasteboardReadingOptionKey = NSString;
@@ -116,7 +116,8 @@ impl NSPasteboard {
     pub fn set_text(&self, text: Id<NSString, Shared>) {
         let _: NSInteger = unsafe { msg_send![self, clearContents] };
         let string_array = NSArray::from_slice(&[text]);
-        if !unsafe { msg_send_bool![self, writeObjects: &*string_array] } {
+        let res: bool = unsafe { msg_send![self, writeObjects: &*string_array] };
+        if !res {
             panic!("Failed writing to pasteboard");
         }
     }